pyproject.toml
[tool.poetry]
name = "coworker"
version = "0.1.0"
description = "AI Coworker"
authors = ["Raza Habib <raza@humanloop.com>", "Jordan Burgess <jordan@humanloop.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
slack-bolt = "^1.18.0"
python-dotenv = "^1.0.0"
slack-sdk = "^3.21.3"
humanloop = "^0.7.27"
docstring-parser = "^0.15"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
py-mon = "^1.1.1"
jurigged = "^0.5.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
